requirements.txt
django
django-crispy-forms
Pillow
networkx
matplotlib
Pandas
django_heroku
gunicorn